E2e Test: add whitespace & remove redundancy
Yet another e2e cleanup. Again with to goal to increased readability and focus.
IMHO the check if the `send` was successful ( or Truthy) is redundant. 'cause if not, all later asserts would fail anyway. Same with array length of requests - unless its part of the actual assertions.
To clearly indicate the test structure of setup,execution,assertion, each section is now separated with an empty line.